Move instantiation and discovery logic to mixin
