+++ npc_engine/services/base_service.py
@@ -1,34 +1,20 @@
 """Module with Model base class."""
 from abc import ABC, abstractmethod
-import os
 from typing import Callable, Dict
-import yaml
+import os
 import zmq
 from loguru import logger
 from jsonrpc import JSONRPCResponseManager, Dispatcher
 from pathlib import Path
 from npc_engine.service_clients.control_client import ControlClient
+from npc_engine.services.factory_mixin import FactoryMixin
 
-from npc_engine.services.utils.config import get_type_from_dict
 
-
-class BaseService(ABC):
+class BaseService(FactoryMixin, ABC):
     """Abstract base class for managed services."""
 
-    models = {}
-
-    def __init_subclass__(cls, **kwargs):
-        """Init subclass where service classes get registered to be discovered."""
-        super().__init_subclass__(**kwargs)
-        cls.models[cls.__name__] = cls
-
     def __init__(
-        self,
-        service_id: str,
-        context: zmq.Context,
-        uri: str,
-        *args,
-        **kwargs,
+        self, service_id: str, context: zmq.Context, uri: str, *args, **kwargs,
     ):
         """Initialize the service.
 
@@ -37,6 +23,7 @@ class BaseService(ABC):
             uri (str): URI to serve requests to
             dependency_clients (list(ServiceClient)): List of dependency clients
         """
+        super(BaseService, self).__init__()
         self.zmq_context = context
         self.socket = context.socket(zmq.REP)
         self.socket.setsockopt(zmq.LINGER, 0)
@@ -52,25 +39,6 @@ class BaseService(ABC):
     def get_api_name(cls) -> str:
         """Return the name of the API."""
         pass
-
-    @classmethod
-    def create(cls, context: zmq.Context, path: str, uri: str, service_id: str):
-        """Create a service from the path.
-
-        Args:
-            context (zmq.Context): ZMQ context
-            path (str): Path to the service
-            uri (str): URI to serve requests to
-
-        Returns:
-            Service: Service instance
-        """
-        config_path = os.path.join(path, "config.yml")
-        with open(config_path) as f:
-            config_dict = yaml.load(f, Loader=yaml.Loader)
-        config_dict["model_path"] = path
-        model_cls = cls.models[get_type_from_dict(config_dict)]
-        return model_cls(**config_dict, context=context, uri=uri, service_id=service_id)
 
     def create_client(self, name: str):
         """Get a dependency client by name to use it in service logic.
@@ -89,7 +57,7 @@ class BaseService(ABC):
         api_name = self.control_client.get_service_metadata(name)["api_name"]
         return ControlClient.get_api_client(api_name)(self.zmq_context, name)
 
-    def start(self):
+    def loop(self):
         """Run service main loop that accepts json rpc."""
         try:
             dispatcher = Dispatcher()
